Build from appropriate dist-git branch
The rawhide branch is not always safe for test rebuilds.
Active development is happening there, and so the bits
sometimes fail to build, causing this test to report false
positives.

Specifically for rawhide kernel, it's not unusual to fail
building during the two week upstream merge window, I've
seen all arches in x86 cause issues during the merge window.
Outside of that things tend to be OK.