Switch to comparative test mode using a baseline
This update switches to a "comparative" mode.  It sets up a baseline by
rebuilding the components used for testing in a clean side tag not
having the component under test in it.  After that it will tag the
component under test to the side tag, run the rebuilds again and compare.